Add FilteredList to ListSequence. Fix count on EntrySequence and BiSequence.

+++ src/main/java/org/d2ab/sequence/EntrySequence.java
@@ -1098,7 +1098,7 @@ public interface EntrySequence<K, V> extends Iterable<Entry<K, V>> {
 	/**
 	 * @return the count of elements in this {@code EntrySequence}.
 	 */
-	default int count() {
+	default long count() {
 		return Iterables.count(this);
 	}
